Detail page api fix

+++ app/src/main/java/com/example/stonks/data/AnalysisResponse.kt
@@ -16,5 +16,5 @@ data class Trend(
 
 data class Growth(
     @SerializedName("raw")
-    val value: Long
+    val value: Float
 )
